Misc.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@42287 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/StatusArg.cpp
@@ -133,7 +133,7 @@ bool StatusVector::ImplStatusVector::append(const ISC_STATUS* const from, const 
 	m_length += copied;
 	m_status_vector[m_length] = isc_arg_end;
 
-	return copied == count;
+	return copied == static_cast<unsigned int>(count);
 }
 
 void StatusVector::ImplStatusVector::shiftLeft(const Base& arg) throw()
